requirements.txt
requests
requests[socks]
ffmpy3
ffmpeg-python
brotli
websockets
websocket-client